Rename Step4_ChoosePlanView to ChoosePlanView
