--- app/ghostery-browser-hub/Views/OnboardingViews/Step4_ChoosePlanView/index.js
@@ -1,16 +1,0 @@
-/**
- * Point of entry index.js file for Ghostery Browser Hub Choose Plan View
- *
- * Ghostery Browser Extension
- * https://www.ghostery.com/
- *
- * Copyright 2020 Ghostery, Inc. All rights reserved.
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0
- */
-
-import ChoosePlanView from './ChoosePlanView';
-
-export default ChoosePlanView;
