+++ app/scss/hub.scss
@@ -87,7 +87,7 @@ html, body, #root {
 @import '../hub/Views/CreateAccountView/CreateAccountView.scss';
 @import '../hub/Views/UpgradePlanView/UpgradePlanView.scss';
 @import '../ghostery-browser-hub/Views/OnboardingViews/Step0_WelcomeView/WelcomeView.scss';
-@import '../ghostery-browser-hub/Views/OnboardingViews/Step4_ChoosePlanView/Step4_ChoosePlanView.scss';
+@import '../ghostery-browser-hub/Views/OnboardingViews/ChoosePlanView/ChoosePlanView.scss';
 @import '../ghostery-browser-hub/Views/OnboardingViews/Step5_SuccessView/SuccessView.scss';
 
 // Imports from ../shared-components directory
